requirements.txt
langchain
langchain-core
langgraph
langchain-groq
langchain-opentutorial
python-dotenv
pydantic
numpy
pandas
matplotlib
yfinance
websockets
gradio
IPython
langchain-mcp